only keep node/

+++ node/deps/v8/src/ast/context-slot-cache.h
@@ -38,7 +38,7 @@ class ContextSlotCache {
     for (int i = 0; i < kLength; ++i) {
       keys_[i].data = NULL;
       keys_[i].name = NULL;
-      values_[i] = static_cast<uint32_t>(kNotFound);
+      values_[i] = kNotFound;
     }
   }
